fix mouse data

+++ src/libs/include/kernel/key.h
@@ -32,7 +32,7 @@ typedef struct keyboard_data {
 typedef struct mouse_data {
     uint8_t flags, z, id;
     bool left, right, mid, btm, top;
-    int32_t x, y;
+    int16_t x, y;
     sq_queue *queue;
     uint32_t info_flag;
 } mouse_data;
